[air][tune] replace node:<ip> custom resource with NodeAffinitySchedulingPolicy
