+++ python/ray/tune/progress_reporter.py
@@ -16,6 +16,7 @@ import pandas as pd
 
 import ray
 from ray._private.dict import flatten_dict
+from ray.air.util.node import _force_on_current_node
 from ray.tune.callback import Callback
 from ray.tune.logger import pretty_print
 from ray.tune.result import (
@@ -36,7 +37,6 @@ from ray.tune.result import (
 from ray.tune.experiment.trial import DEBUG_PRINT_INTERVAL, Trial, _Location
 from ray.tune.trainable import Trainable
 from ray.tune.utils import unflattened_lookup
-from ray.tune.utils.node import _force_on_current_node
 from ray.tune.utils.log import Verbosity, has_verbosity, set_verbosity
 from ray.util.annotations import DeveloperAPI, PublicAPI
 from ray.util.queue import Empty, Queue
