+++ python/ray/tune/utils/file_transfer.py
@@ -8,8 +8,9 @@ from typing import Optional, Tuple, Dict, Generator, Union, List
 
 import ray
 from ray.util.annotations import DeveloperAPI
+from ray.util.scheduling_strategies import NodeAffinitySchedulingStrategy
 from ray.air._internal.filelock import TempFileLock
-from ray.util.scheduling_strategies import PlacementGroupSchedulingStrategy
+from ray.air.util.node import _get_node_id_from_node_ip
 
 
 _DEFAULT_CHUNK_SIZE_BYTES = 500 * 1024 * 1024  # 500 MiB
@@ -114,10 +115,15 @@ def _sync_dir_on_same_node(
         None, or future of the copy task.
 
     """
+
+    node_id = _get_node_id_from_node_ip(ip)
+
     copy_on_node = _remote_copy_dir.options(
         num_cpus=0,
-        resources={f"node:{ip}": 0.01},
-        scheduling_strategy=PlacementGroupSchedulingStrategy(placement_group=None),
+        scheduling_strategy=NodeAffinitySchedulingStrategy(
+            node_id=node_id,
+            soft=False,
+        ),
     )
     copy_future = copy_on_node.remote(
         source_dir=source_path, target_dir=target_path, exclude=exclude
@@ -165,15 +171,23 @@ def _sync_dir_between_different_nodes(
         None, or Tuple of unpack future, pack actor, and files_stats future.
 
     """
+
+    source_node_id = _get_node_id_from_node_ip(source_ip)
+    target_node_id = _get_node_id_from_node_ip(target_ip)
+
     pack_actor_on_source_node = _PackActor.options(
         num_cpus=0,
-        resources={f"node:{source_ip}": 0.01},
-        scheduling_strategy=PlacementGroupSchedulingStrategy(placement_group=None),
+        scheduling_strategy=NodeAffinitySchedulingStrategy(
+            node_id=source_node_id,
+            soft=False,
+        ),
     )
     unpack_on_target_node = _unpack_from_actor.options(
         num_cpus=0,
-        resources={f"node:{target_ip}": 0.01},
-        scheduling_strategy=PlacementGroupSchedulingStrategy(placement_group=None),
+        scheduling_strategy=NodeAffinitySchedulingStrategy(
+            node_id=target_node_id,
+            soft=False,
+        ),
     )
 
     if force_all:
@@ -181,8 +195,10 @@ def _sync_dir_between_different_nodes(
     else:
         files_stats = _remote_get_recursive_files_and_stats.options(
             num_cpus=0,
-            resources={f"node:{target_ip}": 0.01},
-            scheduling_strategy=PlacementGroupSchedulingStrategy(placement_group=None),
+            scheduling_strategy=NodeAffinitySchedulingStrategy(
+                node_id=target_node_id,
+                soft=False,
+            ),
         ).remote(target_path)
 
     pack_actor = pack_actor_on_source_node.remote(
@@ -217,10 +233,15 @@ def delete_on_node(
         Boolean indicating if deletion succeeded, or Ray future
         for scheduled delete task.
     """
+
+    node_id = _get_node_id_from_node_ip(node_ip)
+
     delete_task = _remote_delete_path.options(
         num_cpus=0,
-        resources={f"node:{node_ip}": 0.01},
-        scheduling_strategy=PlacementGroupSchedulingStrategy(placement_group=None),
+        scheduling_strategy=NodeAffinitySchedulingStrategy(
+            node_id=node_id,
+            soft=False,
+        ),
     )
     future = delete_task.remote(path)
